+++ docker/scripts/install_ldak.sh
@@ -1,6 +1,6 @@
 #!/bin/sh
 set -euo pipefail
 
-wget --no-check-certificate https://dougspeed.com/wp-content/uploads/ldak5.2.linux_.zip
-unzip ldak5.2.linux_.zip
-cp ldak5.2.linux /bin/ldak
+wget https://github.com/dougspeed/LDAK/blob/4ee871be17d8ea406494211638a5ead677e7dd47/ldak6.linux
+chmod +x ldak6.linux
+cp ldak6.linux /bin/ldak